requirements.txt
colorlog
python-dotenv
Flask
flask-cors
flask-marshmallow
Flask-Migrate
flask-restful
Flask-SQLAlchemy
marshmallow-sqlalchemy
PyJWT
requests
